+++ libzebra/src/main/cpp/include/libavutil/murmur3.h
@@ -27,7 +27,6 @@
 #ifndef AVUTIL_MURMUR3_H
 #define AVUTIL_MURMUR3_H
 
-#include <stddef.h>
 #include <stdint.h>
 
 /**
@@ -62,7 +61,7 @@
 /**
  * Allocate an AVMurMur3 hash context.
  *
- * @return Uninitialized hash context or `NULL` in case of error
+ * @return Uninitialized hash context or `nullptr` in case of error
  */
 struct AVMurMur3 *av_murmur3_alloc(void);
 
@@ -98,7 +97,7 @@ void av_murmur3_init(struct AVMurMur3 *c);
  * @param[in]  src  Input data to update hash with
  * @param[in]  len  Number of bytes to read from `src`
  */
-void av_murmur3_update(struct AVMurMur3 *c, const uint8_t *src, size_t len);
+void av_murmur3_update(struct AVMurMur3 *c, const uint8_t *src, int len);
 
 /**
  * Finish hashing and output digest value.
